MINI2440: Add machine support
The MINI2440 is a chinese made s3c2440 development board with
a large set of peripherals.
This patch provides machine support for almost all the features
of the board.

Since it can come with various "options" fitted, a kernel parameter
is used to specify the lcd size, backlight control and touchscreen.

Signed-off-by: Michel Pollet <buserror@gmail.com>