s3c-fb: udpate to support s3c2416/s3c2443 style hardware
Update the variant and window variant structures with the necessary
changes to support the older style of hardware where these are not
in the same place.

Add the support for the s3c2443/s3c2416 hardware by using the
platform-device s3c2443 to cover both, and add the initialisation
data for these.

Also change to including just the v4 header files for the moment
until the last of the merging of these is sorted out.

Signed-off-by: Ben Dooks <ben-linux@fluff.org>