s3c-fb: fixed distortedness situation for the mode more then 24bpp.
it has been working fine at 16bpp but in case of pixel format more then
24bpp
it would occur distortedness situation on that mode. so this patch set
the word swap control bit of WINCONx to 1 as default value. but it
should be set to 0
in case that each ENLOCAL bit of WINCON0 ~ 2 registers is enabled.
this issue would be solved with local path feature soon.

Signed-off-by: InKi Dae <inki.dae@samsung.com <mailto:inki.dae@samsung.com>>
Reviewed-by: KyungMin Park <kyungmin.park.samsung.com
<http://kyungmin.park.samsung.com>>
Signed-off-by: Ben Dooks <ben-linux@fluff.org>